Break out error components into separate files; more lint
