+++ .eslintrc
@@ -431,6 +431,38 @@
             2,
             "tag-aligned"
         ],
+        "react/jsx-curly-spacing": [
+            2,
+            "never"
+        ],
+        "react/jsx-indent-props": [
+            2,
+            2
+        ],
+        "react/jsx-no-duplicate-props": [
+            2
+        ],
+        "react/jsx-no-undef": [
+            2
+        ],
+        "react/jsx-uses-react": [
+            2
+        ],
+        "react/jsx-uses-vars": [
+            2
+        ],
+        "react/no-did-mount-set-state": [
+            2
+        ],
+        "react/no-did-update-set-state": [
+            2
+        ],
+        "react/no-direct-mutation-state": [
+            2
+        ],
+        "react/no-multi-comp": [
+            2
+        ],
         "quote-props": [
             2,
             "as-needed",
